fix: resolve Clippy warnings in vx-output-filter
